Add a copy of boost library in order to be able to include header-only parts of it
that way we are building only the used header-only libs and there is no need for installation.
We are not going to modify it, the updates should be easy. Yes, it's a lot of files, we can strip them to the couple we are using but it's not a problem for now