Add a little bash shell script for measuring the time for starting and running an executable file
Basically, the script runs the provided executable (also, forwards the rest of the arguments to it) as many times as it can in 10 seconds and prints the number of runs and average time per run.
In order to double-check that your executable has been indeed run the output of the last run is written to a txt file (check the script's logging messages).

An example of usage can be found in the script.