[optimized_measured_model] Add a new optimized implementation for a measured model
See measured model for more info regarding the idea of the model.
This one finds a satisfiable model with the faster algorithm(via the 'model') and adds additional
points for the atomic measured formulas (one for each formula).
Let's the number of additonal points is A and the number used variables is N.
All different unique evaluations are 2^N.
We need a subset of size A from thouse evaluations for the additional points (we can have duplications).
So, all combinations of such subsets are (2^N)^A or 2^(N*A).

Compared to the 'measured_model' where it is roufly 2^(N*(A+U)), where U is the number atomic contacts * 2 + number of atomic !=0.
It is a sagnificant optimization.
Now, this measured model is 'bounded' from the number of atomic measured formulas instead of also the
number of atomic contacts and !=0 formulas.