[rest_server] Update the client accumulated logging output also in the termination callback
a little ugly but the accumulated output should be updated more often, not only when printing something,
because there might be huge intervals in which the algorithm does not print anything and some of the already accumulated log info will not be pushed to the final_result's output
till a new log message is received.
We want to provide all the accumulated output to that point so when there is no new output the user will know where the algorithm is slow.