[rest_server] Fix bug with a not updating output when there are multiple tasks
the use case was: start of measured model task, then start a fast model task, the fast model was not updating it's output until the measured model task was not stopped.
Note that with two measured model tasks it was working fine, most probably because the measured model checks for termination
(which updates also the output) more frequently than the fast model.
Nevertheless, it's a bug because a static variable in lambda function is the same for all function calls(i.e. for all tasks),
there is one task per thread but the variable is static, so it was even a race condition (multiple threads were trying to read/write to it)