[formula_lib] Change the term evaluating with a variable evaluation to not create subterms
If the variable evaluation does not evaluate some of the boolean variables in the term it's going to create a subterm with the not evaluated portions.
Currently, there is no need for a term evaluation with a not full variable evaluation.
So, by default it will not make a subterm. All models are passing full variable evaluations so it will not break them.